test(OMN-61 phase 1): fix latent vacuity in OMN-47 parity + add reverse guard
Code review (PR #11) found the new file (a) redundant with the existing
OMN-47 schema-impl-parity forward parity tests and (b) vacuous for project
name/flagged/dueDate — the generated script's nextTask block emits those keys
unconditionally, so `\bfield:` matched even with the projection case deleted.
The pre-existing OMN-47 tests shared the same flaw: the parity net meant to
catch field drift has itself been silently vacuous for 3 fields (an OMN-61
meta-instance — the safety mechanism had the gap it was built to prevent).

- schema-impl-parity.test.ts: both forward parity loops now build with
  performanceMode:'lite' (suppresses nextTask/taskCounts) — de-vacuumed.
  Proven: removing the project `name` projection case now FAILS the OMN-47
  ProjectFieldEnum parity test (previously a false pass).
- Added the reverse-direction guard (no undeclared top-level projection key,
  context-only allowlist) into the OMN-47 file per its stated convention.
- Deleted the redundant field-projection-parity.test.ts.

Full unit suite 1826/1826; build, prettier, eslint clean.
